refactor: migrate write-offs to useTransferTableYear, move create dialog up to view
