+++ apps/dashboard/src/modules/financial/components/write-offs/WriteOffTable.vue
@@ -13,15 +13,6 @@
       :value="writeOffs"
       @page="onPage"
     >
-      <template #header>
-        <div class="flex flex-row align-items-center justify-content-between">
-          <IconField icon-position="left">
-            <InputIcon class="pi pi-search" />
-            <InputText v-model="idQuery" :placeholder="t('common.id')" @focusout="searchId" @keyup.enter="searchId" />
-          </IconField>
-          <Button class="ml-2" icon="pi pi-plus" :label="t('common.create')" @click="showDialog = true" />
-        </div>
-      </template>
       <Column field="id" :header="t('common.id')">
         <template #body="slotProps">
           <Skeleton v-if="isLoading" class="w-6 my-1 h-1rem surface-300" />
@@ -81,18 +72,6 @@
     >
       {{ t('modules.financial.write-offs.delete.not-possible') }}
     </Dialog>
-
-    <FormDialog
-      v-model="showDialog"
-      :confirm="true"
-      :form="form"
-      :header="t('modules.financial.write-offs.create')"
-      :is-editable="true"
-    >
-      <template #form="slotProps">
-        <WriteOffCreateForm :form="slotProps.form" @submit:success="showDialog = false" />
-      </template>
-    </FormDialog>
   </div>
 </template>
 
@@ -107,11 +86,7 @@ import { useI18n } from 'vue-i18n';
 import { addListenerOnDialogueOverlay } from '@sudosos/sudosos-frontend-common';
 import { useToast } from 'primevue/usetoast';
 import type { DataTablePageEvent } from 'primevue/datatable';
-import FormDialog from '@/components/FormDialog.vue';
-import WriteOffCreateForm from '@/modules/financial/components/write-offs/forms/WriteOffCreateForm.vue';
 import { formatDateFromString, formatPrice } from '@/utils/formatterUtils';
-import { schemaToForm } from '@/utils/formUtils';
-import { createWriteOffSchema } from '@/utils/validation-schema';
 import { getWriteOffPdfSrc } from '@/utils/urlUtils';
 import { useWriteOffStore } from '@/stores/writeoff.store';
 
@@ -128,7 +103,7 @@ withDefaults(
   },
 );
 
-const emit = defineEmits(['page', 'single']);
+const emit = defineEmits(['page']);
 
 const { t } = useI18n();
 const writeOffStore = useWriteOffStore();
@@ -140,18 +115,10 @@ const showWarning = () => {
   showWarningModal.value = true;
 };
 
-const showDialog = ref(false);
-const form = schemaToForm(createWriteOffSchema);
-
-const idQuery = ref('');
 const downloadingPdf = ref(false);
 
 function onPage(event: DataTablePageEvent) {
   emit('page', event);
-}
-
-function searchId() {
-  emit('single', parseInt(idQuery.value));
 }
 
 const getName = (writeOff: WriteOffResponse) => {
